services/backend/requirements.txt
python-dotenv==1.0.1
pandas==2.2.1
pydantic==2.10.3
python-multipart==0.0.19

# Redis
redis==5.0.5

# Text processors
pymupdf==1.24.12
pymupdf4llm==0.0.17


# Embeddings
sentence-transformers==3.3.1
langchain-huggingface==0.1.2

# LLM
tiktoken==0.8.0
langchain==0.3.10
langchain-community==0.3.10
langchain-text-splitters==0.3.2
#transformers==4.46.3

# OpenAI
langchain-openai==0.2.11
openai==1.57.0

# Google
langchain_google_genai==2.0.5

# Back-end
fastapi==0.105.0
uvicorn==0.24.0